Basic support for using NPM and Node to manage webapps
This allows us to remove Maven from the frontend workflows, instead
publishing frontend applications directly to NPM.

The `dhis-web-apps` maven project then executes one command with NPM and
one with Node.

1. `npm install`
2. 'node ./copy-apps.js'

The first installs the dependencies to `node_modules`. In this case all
our applications would be published to NPM.

The second copies the apps from `node_modules/<app>` to
`target/dhis-web-apps/<app>` just like the previous JAR-files where
unpacked to.

There are some benefits to this as I have outlined in:
https://docs.google.com/document/d/1DTY22D1CDpPQqqFOENkJKhHyhgTT6j8XRgARo84TtwY/edit#heading=h.fxuxtg782lfg

The major ones are:

1. Continuous deployment of web apps
2. Simplified workflow for app developers
3. The ability to deploy beta-versions, canary release, deprecate versions, etc.
4. The ability to freeze versions in the WAR-file
5. The ability to create custom WAR-file builds with any apps by just
   changing the package.json file.
6. Reproducible builds and knowledge about what code is running in what
   build